Moved file upload to posts controller
